// frontend/generated/models/ProjectDto.ts
export interface ProjectDto {
    /**
     * 
     * @type {number}
     * @memberof ProjectDto
     */
    projectId?: number;
    /**
     * 
     * @type {ProjectCategoryDto}
     * @memberof ProjectDto
     */
    projectCategory?: ProjectCategoryDto;
    /**
     * 
     * @type {AccountDto}
     * @memberof ProjectDto
     */
    client?: AccountDto;
    /**
     * 
     * @type {AccountDto}
     * @memberof ProjectDto
     */
    staff?: AccountDto;
    /**
     * 
     * @type {string}
     * @memberof ProjectDto
     */
    title?: string;
    /**
     * 
     * @type {string}
     * @memberof ProjectDto
     */
    description?: string;
    /**
     * 
     * @type {ProjectStatusDto}
     * @memberof ProjectDto
     */
    status?: ProjectStatusDto;
    /**
     * 
     * @type {Date}
     * @memberof ProjectDto
     */
    startDate?: Date;
    /**
     * 
     * @type {number}
     * @memberof ProjectDto
     */
    minBudget?: number;
    /**
     * 
     * @type {number}
     * @memberof ProjectDto
     */
    maxBudget?: number;
    /**
     * 
     * @type {Date}
     * @memberof ProjectDto
     */
    createdAt?: Date;
    /**
     * 
     * @type {Date}
     * @memberof ProjectDto
     */
    updatedAt?: Date;
    /**
     * 
     * @type {boolean}
     * @memberof ProjectDto
     */
    isVisible?: boolean;
    /**
     * 
     * @type {Array<ReportDto>}
     * @memberof ProjectDto
     */
    reports?: Array<ReportDto>;
    /**
     * 
     * @type {Array<FileDto>}
     * @memberof ProjectDto
     */
    files?: Array<FileDto>;
    /**
     * 
     * @type {ContractDto}
     * @memberof ProjectDto
     */
    contract?: ContractDto;
    /**
     * 
     * @type {Array<MilestoneDto>}
     * @memberof ProjectDto
     */
    milestones?: Array<MilestoneDto>;
    /**
     * 
     * @type {MilestoneDto}
     * @memberof ProjectDto
     */
    activeMilestone?: MilestoneDto;
    /**
     * 
     * @type {Array<ProjectSkillDto>}
     * @memberof ProjectDto
     */
    requiredSkills?: Array<ProjectSkillDto>;
    /**
     * 
     * @type {number}
     * @memberof ProjectDto
     */
    proposalCount?: number;
    /**
     * 
     * @type {MessageDto}
     * @memberof ProjectDto
     */
    latestMessage?: MessageDto;
    /**
     * 
     * @type {ProjectTerminationReasonDto}
     * @memberof ProjectDto
     */
    terminationReason?: ProjectTerminationReasonDto;
    /**
     * 
     * @type {boolean}
     * @memberof ProjectDto
     */
    toTerminate?: boolean;
    /**
     * 
     * @type {Array<ProposalDto>}
     * @memberof ProjectDto
     */
    myProposals?: Array<ProposalDto>;
}